Update tests to use new functions
