+++ server/planning/__init__.py
@@ -48,7 +48,6 @@ from planning.events import init_app as init_events_app
 from planning.planning import init_app as init_planning_app
 from planning.assignments import init_app as init_assignments_app
 from planning.search import init_app as init_search_app
-from planning.validate import init_app as init_validator_app
 from planning.locations import init_app as init_locations_app
 from superdesk.celery_app import celery
 from .published_planning import PublishedPlanningResource, PublishedPlanningService
@@ -115,7 +114,6 @@ def init_app(app):
     init_planning_app(app)
     init_assignments_app(app)
     init_search_app(app)
-    init_validator_app(app)
     init_planning_download_app(app)
     init_planning_locks_app(app)
     init_planning_autocomplete_app(app)
